Split up assembler code into multiple files.
