+++ cpp/dolfin/fem/CMakeLists.txt
@@ -1,5 +1,7 @@
 set(HEADERS
   assembler.h
+  assemble_matrix_impl.h
+  assemble_vector_impl.h
   AssemblerBase.h
   CoordinateMapping.h
   DirichletBC.h
@@ -23,6 +25,8 @@ set(HEADERS
 
 set(SOURCES
   assembler.cpp
+  assemble_matrix_impl.cpp
+  assemble_vector_impl.cpp
   AssemblerBase.cpp
   DirichletBC.cpp
   DiscreteOperators.cpp
